+++ reana_server/config.py
@@ -183,12 +183,5 @@ ADMIN_EMAIL = os.getenv("REANA_EMAIL_SENDER", "CHANGE_ME")
 
 # Workflow scheduler
 # ==================
-REANA_SCHEDULER_SECONDS_TO_WAIT_FOR_REANA_READY = int(
-    os.getenv("REANA_SCHEDULER_SECONDS_TO_WAIT_FOR_REANA_READY", "600")
-)
-"""How many seconds to wait between retries in case of REANA not ready to run more workflows."""
-
-REANA_SCHEDULER_SECONDS_RETRY_DELAY = int(
-    os.getenv("REANA_SCHEDULER_SECONDS_RETRY_DELAY", "60")
-)
-"""How many seconds for a workflow to be delayed once rescheduled."""
+REANA_SCHEDULER_REQUEUE_SLEEP = int(os.getenv("REANA_SCHEDULER_REQUEUE_SLEEP", "15"))
+"""How many seconds to wait between consuming workflows."""